Level design preparations

+++ Assets/Scripts/Game/SO/UnitSO.cs
@@ -8,7 +8,7 @@ public class UnitSO : ScriptableObject
     public UnitTypeSO UnitType => _unitType;
     public int MaxHealth => _maxHealth;
     public int Damage => _damage;
-    public int EnemyDetectionDistance => _enemyDetecationDistance;
+    public float EnemyDetectionDistance => _enemyDetecationDistance;
     public float Speed => _speed;
     public float AttackRate => _attackRate;
     public float AttackRange => _attackRange;
@@ -18,7 +18,7 @@ public class UnitSO : ScriptableObject
     [SerializeField] private UnitComp _prefab;
     [SerializeField] private UnitTypeSO _unitType;
     [SerializeField] private int _maxHealth;
-    [SerializeField] private int _enemyDetecationDistance;
+    [SerializeField] private float _enemyDetecationDistance;
     [SerializeField] private int _damage;
     [SerializeField] private float _speed;
     [SerializeField] private float _attackRate = 0.5f;
